Deep clone expression result before modifying
Necessary for the follow up which tickled the bug.  The cached result is
mutated in the action, followed by some backtracking so that when the
action is hit again, the mutated result is used which drops some tokens
from the pushed attribute.

In order to prevent similar types of bugs, while running parserTests the
results placed in the tokenizer cache are now frozen to catch any
attempts at mutating them.  Verified that this would have caught the
above manually located bug and, indeed, it susses out another location
where mutation occurs.

Change-Id: I380bb7a9bb16026954eb39a24c5915f5992153a9