Replace callbacks with Promises in Logger
The log method now returns a Promise which will be fulfilled when the
log operation is completed on all backends.  In most cases this can be
ignored, but it can be used (for example) to ensure that parse waits
for the linter backend to flush its output before returning.

This also removes our use of the `async` module in lib/logger.

Logger-related classes removed from jsduck configuration due to T156469.
Noted in T184573 to re-add it once jsduck is fixed/replaced.

Change-Id: I6e6780360f3b44ba2f08dd1d131ad7119f4c9a60