Use Promise.async/yield in WTS._serializeAttributes
Use of Promise.reduce was rewritten to use a `for of` loop over an
array, which is supported in node >= 4.  Declarations using `var` were
rewritten to use `let` or (where eslint forced us to) `const`, in
order to manage the name conflicts caused by the change of scoping,
now that we're not creating a bunch of new function scopes.

Change-Id: Ib8404bc5416738a84ded404e568be507c5d3c1ba