Remove fixme in handlepres
The php parser is quirky.  Even though div, blockquote, and center
induce closing a paragraph, because their opening tags match the regexp
for closeMatch (/?), they don't move to the inBlockElem state, and hence
indent pre is permissible inside them.  There's a special case for
blockquote however, with inBlockquote, which results in that being the
exception of the three.

https://github.com/wikimedia/mediawiki/blob/4b2f95bed5a9ae0a917264abed2ba1fd9fb78f3c/includes/parser/BlockLevelPass.php#L323

Bug: T53262
Change-Id: I37564ee28897b16cfd99caa8a2eca54842e2148e