html2wt: Add new lines before and after lists in wikitext
This patch adds newlines around lists to make them
standout when viewed as wikitext. Selser will preserve
the newline absense in existing wikitext.

Wrote a series of selser tests to validate proper behavior.

A new "wrap" command was added to the parserTest.js
file to enable a specific new type of selser test
to be written where a list is wrapped in a table cell.

A bunch of tests are failing wt2wt mode but the tests
shouldn't be fixed up because the tests require the
absence of newlines around those lists.

Bug: T177102
Change-Id: Icd61dc3a0d7ae686e07dd0aacf4aeeadcc89f064