Prevent crashing on foundationwiki pages
This is the downside of the optimization added in d47d5ab.  If an
extension tag returns elements with invalid attributes, when doing the
adoption we now throw.

For example,
http://localhost:8000/wikimediafoundation.org/v3/page/html/L11_ExtraInfo_0902_Y%2Fen%2FUS/67456

See the following for a typical use of the "html" extension tag,
https://wikimediafoundation.org/w/index.php?title=Template:2011FR/lp-form-US7amounts-extrainfo&action=edit

A less expensive fix would be to implement T179082

Bug: T185266
Change-Id: I57d146e20793d7229a0a3944db7e332525c72133