Normalize away unnecessary attributes in data-mw.html too
"about" attributes on elements in data-mw.html are another source of
indeterminacy, as noted in T93715, but were being missed because of the
JSON encoding not matching the regexp.

The new errors on,

node bin/roundtrip-test.js --domain en.wikipedia.org "Nicolas Iljine"

were being classified as semantic when the numbers didn't line up.

Bug: T151474
Change-Id: I0b906b588a983030d49ca361141fcc1e77c4b452