Make the Sanitizer "static" and decouple it from the parsing pipeline
We have had to do this a few times now when extensions and some
other parts of code have had to access sanitizer functionality
and cannot because we have tied it to the parsing pipeline.

The SanitizerHandler object provides that parsing pipeline hook
and lets all other parsing functionality treat the sanitizer
as a purely functional interface without any internal state.

Change-Id: I113331f7b34c37dc300f4916d00f70fe79cc89f7