Revert "Fix inefficiency in SyncTokenTransformManager"
This reverts commit c19dab6b383f190946b9bf3c3b9c4b2d9753ac82.

This exposed a latent bug in the token transformers code where
token handlers can modify the parsing pipeline by adding/removing
handlers. That bug is only exercised very rarely, it appears, and looks like
the reverted patch is one of those scenarios. Till we fix that latent bug,
it is simpler to revert this patch.

Bug: T191843
Change-Id: Ic5c20c3e97399b5c61f698eff4084acc10a15c50