Unconditionally strip section tags
The information on it is redundant and is in no case necessary for
serialization.  This matches the implementation in VE.

See I43ced131d07a20318af37e830318e30db2eb451f

Follow up to f61cbce

The fix there was necessary but not sufficient.  See the difference
with this change on,

http://localhost:8000/_rt/mediawikiwiki/Manual%3AHooks
http://localhost:8000/_rtselser/mediawikiwiki/Manual%3AHooks

Bug: T192909
Change-Id: I3394dac36363bca0e4468a15460567e575069e55