Use lowercase names in jsdoc for primitive types.
The primitive type `string` is actually different from an instance of a
`String` object in JavaScript; be consistent in using the lowercased
name for the primitive object when that's what we mean.  Our codebase
is mostly consistent about this; dunno how these exceptions crept in,
but it was probably my fault.

Change-Id: I0d9ed4b4a0ec2b015da7bdae480e93674610038b