Implement pb2pb endpoint for language variant conversion
We also split up env.htmlVariantLanguage / env.wtVariantLanguage, which
lets us distinguish between various cases when round-tripping (which
isn't implemented yet).

In the process, added a bunch of test cases for the wt2html endpoint
as well and ensured that we return an HTTP Vary header to indicate to
RESTBase where its storage needs to take variants into account.

Change-Id: Idc4d16bda729a50f08c9ec7ffcbc76888a9cfd15