Move to DOM based wrapping of bare text
This effectively reverts 3d53f56 since the proper fix is probably to get
Remex to account for fostered content.  Otherwise, we end up with
p-wrapped foster spans.

dt/dd/li are added to the always suppressing tags to match ce9f5c2 in
core.

The cleanupFormattingTagFixup pass is removed since its job was to move
formatting tags that ended up *inside* figures into figcaptions.  Since
that no longer happens, it's a no-op.  It was introduced in T65642.

However, in T68749, there's a request/argument to avoid nesting blocks
in formatting tags, since VE can't handle that data model violation and
marks the nodes as uneditable.  In the past, we've supported that
inconsistently.  For example, the case in "1. List embedded in a
formatting tag" is uneditable, despite it being a popular pattern.
In "2. Treebuilder fixup of formatting elt", only one of two images was
editable.

We should consider reopening T68749 and adding a pass to solve the issue
generally.  Note that the Parsoid output now matches Remex so any
changes to the DOM is explicitly introducing a difference to support
editing clients, and should be noted as such.

Blacklist changes for "T35845 (2): Headings become bold in TOC when they
contain a blockquote" are because the paragraph in the blockquote
serializes to newlines which are converted to spaces for single-line
context.  Same for "Multiple tags in TOC".  Seemed relevant to leave
these failures for when we develop a better strategy for serializing
newlines found in those nodes.

For various tests, parsoid modes are limited to avoid the need for
specific sections, just to add data-parsoid attributes in order for the
test to roundtrip w/o the normalized form.

Bug: T110004
Change-Id: I5ad254fe5ff8e2b061e2db33288a116a8d88ef05