Effectively revert 50056f6
That doesn't seem like it was the right fix, since it tokenizes the
newline as a string instead of a NlTk.  We've likely fixed the original
bug in other ways since the output is improved with this revert.

The test "Parsoid: Don't paragraph-wrap fosterable content" is removed
since, contrary to what it says, we do p-wrap fosterable content and the
mixed table syntax isn't supported in the php parser.

One of the newly blacklisted selser tests is a false positive, selser is
more accurate.  The other is a case where selser is unnecessarily
preserving a trailing whitespace in a table cell based on the heuristics
added in bd5a840.

Change-Id: I0b5b2366ae0c7ceddbcda04e32a018c4fbe0894d