Respect the proxy overriding even when going through the mwApiServer
In some cases, like private, fishbowl wikis plus wikitech, we skipped
configuring a proxy and connected directly to the original url, when using the
old proxying method. Let's honour the per-site overriding when using internal
HTTPS via mwApiServer.

This should unbreak wikitech (hosted on a different cluster) when we use TLS in
production.

Bug: T163438
Change-Id: If7f148bd3ef5670d31775f880b43b11f69da12cb