Remove strictAcceptCheck config
The config is currently enabled by default and in production.

We added this as a fail-safe a long time ago since we were putting the
content negotiation code in production before coordinating with clients
that we were going to enforce it and didn't want to break anyone
accidentally by returning 406s.

Now that content negotiation has rolled out in earnest, it is preferable
to remove this config because it's a source of confusion if it's disabled
and one is passing Accept headers.  And no feature is lost since clients
can omit the Accept header to receive the latest version of content.

Change-Id: I201474dd3e0fcad1022a4b02303b34df94f81c1f