Don't migrate template start markers past auto-inserted table ends
A follow up to 67277ed as it interacts with 8976ab9

The claim that a flipped range should always be enclosed is somewhat
dubious.  It's guaranteed when a foster box is emitted, since if a
meta marker is found in the foster box, the entire box and table are
wrapped in additional markers.

However, if the table is templated, the foster box isn't emitted.

Here we're in a scenario where a templated table is unclosed and a
subsequent template has its end marker fostered by that table.  In that
case, migrating the start marker outside the table will result in
flipped range that isn't enclosed.  Hence, the fix here to not migrate
out any start markers from tables with auto-inserted ends.

Note that unclosed tables from templates will likely continue to be
thorn in our sides.  For example, try this test case out where the table
doesn't even end up being template marked!

{{1x|
{{{!}}
<div>}}

Fixes the rendering of the following and others,
http://localhost:8000/es.wikipedia.org/v3/page/html/JoJo's_Bizarre_Adventure/110884642
http://localhost:8000/ru.wikipedia.org/v3/page/html/%D0%91%D0%B5%D0%BB%D1%8F%D0%BA%D0%BE%D0%B2%D0%B8%D1%87%2C_%D0%92%D0%B0%D0%BB%D0%B5%D1%80%D0%B8%D0%B9_%D0%A0%D0%BE%D0%BC%D0%B0%D0%BD%D0%BE%D0%B2%D0%B8%D1%87/92951761

The change to the parser test reverts what was made 67277ed

Bug: T205674
Change-Id: Icd3411e559842e647ece3662c7e7205b6f23b919