Fix the logic (and comments) when skipping the funnel into mwApiServer
This is a fixup for If7f148bd3, where I read the comments in
mw/ApiRequest.js incorrectly. If we're funneling requests through a
specific server, we just want to skip the wikis defined as "non-global",
we should not skip the funneling for private and fishbowl wikis.

So, do the following:
- Add the "nonglobal" property from the sitematrix to the api
configuration map
- Add it to the wiki config as well
- Skip funneling requests via the mwApiServer if the wiki is set to
non-global

Please note that for non-global wikis, proxying is skipped anyways as
the proxy uri is forcibly set to "null" in the configuration.

Bug: T163438
Change-Id: I350f17655b3054f00dc606774872c24fdcbb4db1