ConfigRequest doesn't need a separate proxy argument
The concern in d2ef67c about the proxy from the first WikiConfig
constructed before the call to switchToConfig erroneously being used to
make the ConfigRequest was fixed in that commit, as it was instantiated
with the proxy for the supplied prefix.

The refactor in 70fcf9b assures that going forward.

Change-Id: I72d068c4c0db9c10f9751edd876e2942eb0bec79