TokenHandler: move initialization to constructor instead of init() method
Do initialization in the constructor instead of an init() method.
Otherwise #init() gets invoked in the superclass() constructor *before
the child class constructor gets a chance to run* which can be
surprising.  Also this puts field initialization in the constructor
where a reader (or an automated documentation or conversion tool)
would expect to find it.

Change-Id: Ic9db9c3b0c2023f503ad14ca171a8d9704019b36