Separate environment changes per test and mode
The environment is shared between test runs but a new WikiConfig is
constructed by `switchToConfig` in `processTest` which gets shared
across modes.  Some of the setup stays the same for each mode, so only
needs to be done once.  The changes that need to be made per mode are
left in `prepareTest`.

The blacklist changes are because `initializeForPageName` is only called
once now when processing the test, rather than each time a mode is prepared.
That means the environment uid isn't reset between modes, resulting in
higher generated id numbers.

Change-Id: I0266be1969ce795a94e9bdc5b51c88eece8b287c