Stop always using a span representation for inline dom fragments
There's no justification for doing it and it's inconsistent with what we
do for blocks.  Perhaps it harkens back to a time before the code path
was guaranteed to be working on span wrapped tokens.  In any case, we
assert that the nodes are elements.

This avoids the need for special casing style tags in a follow up.

Change-Id: Ica9eace5f45f2dadb7280b33dbb94b65c6920c59