Capture async tokenizing time
Previously, all that was being measured was the construction of the
iterator.  The tokenizing happens on `iterator.next()`.

Note that, because there're recursive calls when tokenizing, time spent
sync tokenzing may be overcounted and async tokenizing can include the
sync time (only *can* because there're other places along the parse
where we do sync tokenzing).

Bug: T206040
Change-Id: Ie8bcaa65f0a62924fce363e48722feec2cbe593a