Handle JSON formatversion=2 for parse/preprocess/batch API requests
Use formatversion=2 for the parse and expandtemplates API requests in
addition to the ParsoidBatchAPI requests.  The ParsoidBatchAPI has
been updated to match the formatversion=2 output of parse and
expandtemplates (I67b8b7459d14b3456f170a595ca085cdd76d9a28), and some
slight tweaks are needed: the category key is now 'category' instead
of '*' and the property list is an object with name/value mappings,
instead of an array of pairs.

Keep the code to parse the old format as well for now, to ease
migration to the updated ParsoidBatchAPI extension.

Change-Id: I8ab01202f700e3f0b87f87ddfd24ce710e668c6d