Added genTest capabilities for DOM passes
--genTest now accepts 21 different DOM dump categories
as part of a new unit test infrastructure. Fragments are
by default disabled in the dumps but can be enabled with
--genTestFragments true command line option.

For dom handler test generation, the output files are
specified by using --pageName along with the
--genDirectory optional argument.

For token transformers --genTestOut ../dumpFileName
supports directory info being specified as part of
the path/filename.

Bug: T209186
Change-Id: I0515f273f237976d2b3bff06f3a1452d92346fa6