Move applying about ids from fragment wrappers to extension content
It's only the extension content that needs the about ids.  This matches
b241ac6 where span wrapping was only applied to extension content, and
goes hand-in-hand.  The one necessitates the other.

Fragment wrappers don't need abouts since they don't have siblings and
are represented as a single node.

When we get to unpacking, it's now clear that if the wrapper has an
about id, it came from template encapsulation and would need to be
applied to the top-level nodes in the fragment.  This being true of
all fragment, it lets use get rid of the `isForeignContent` option to
`encapsulateExpansionHTML`, since that no longer serves to distinguish
anything.

Change-Id: Id9faae3d4e3c8771c2de1fb42ba62a7d92d76673