TemplateHandler:onTemplateArg: Use manager.frame instead of function arg
Frames are associated with pipelines and so new frames are created
when a subpipeline is created for processing content recursively
(ex: templates, extensions, attributes, link dom fragments, other
dom fragments, etc.). Frames are like call-stack entries in a
conventional programming language and contain the arguments of the
extension/template if any. When a new pipeline is constructed (see
PipelineUtils.processContentInPipeline), pipeline.setFrame is called
which constructs a new frame and associates it with the AsyncTTM.

During pipeline construction, all (async) token transformers have a
reference to the AsyncTTM they are associated with. While this makes
the transformers non-functional by associating state and requres them
to be instantiated for every pipeline instance, that has been the case
from the very beginning (as part of parser.js:makePipeline).

So, we can as well leverage that and stop passing around unnecessary
parameters like frame when it is available in this.manager.frame

In followup patches, we'll get rid of the frame arg to the async
token transformers and then eventually unify the token transformer
signatures for the sync and sync transformers.

Change-Id: I49bedd002f2c48a79fd11abee9327f2eccd3f395