Account for span wrapping when fragment is first transclusion node
A regression introduced in b241ac6.

Fixes the crashers on
http://localhost:8000/zh.wikipedia.org/v3/page/html/%E9%8C%B8%E7%9A%84%E5%90%8C%E4%BD%8D%E7%B4%A0/52156747
http://localhost:8000/en.wikipedia.org/v3/page/html/User_talk%3ANetha_Hussain%2FArchive_4/872287243
and others.

An isolated case from the zhwiki one is,
```
[[Test|{{1x|[[Hmm|Something <sup>strange</sup>]]}}]]
```

This edge case of link-in-link exposes that it's possible to end up with
a document fragment at the first encapsulation position of a template
without having span wrapped the fragment contents.  At present, only link
content (ignoring the nowiki extension exception) omits span wrapping
the fragment contents since it's being placed in an anchor.  Normally,
transclusion annotation would end up on the anchor itself, instead of
the link contents, but because of the content model violation here, we
end up trying to place the annotation on the text node ("Something").

This is more than just edge case handling since in the future we may
drop the span wrapping in the nowiki extension or come up with some
other wikitext construct that uses dom fragment without wrapping.

Bug: T211941
Change-Id: Ic398a6082b5f012823afe2c32c4606c70df1c5da