Reduce the batch size for pageprop requests
The default batch size is 50.  However, since redlinks are added
serially in a post-processing pass (T207279), requests call the batch api
directly rather than using the batcher, since they'd otherwise sit idly
waiting for the batcher to make progress.

Although the number of titles in a query was split to match LIMIT_BIG1,
all queries were stuffed in batch, far surpassing the recommended batch
size.

This patch reduces the batch size for these requests to essentially 1,
since that's already getting info for LIMIT_BIG1 number of titles, but
allows for multiple requests to be made in parallel.

Fixes the consistent batching errors on the following and presumably
others.

~30k links
http://localhost:8000/fr.wikipedia.org/v3/page/html/Utilisateur%3AChico75%2Fpages_avec_pourcent/156246032

~40k links
http://localhost:8000/en.wiktionary.org/v3/page/html/User%3APengo%2Fobsolete_genus_list/49976305

Bug: T215537
Change-Id: I4498d5642c9bb08bf0b9dd4e53f482e5cd6e408e