Assert that the .dataobject isn't touched after storing attrs on a node
This is desirable since, in a follow up world where only the index is
present on the node, we want to be able to drop that when storing attrs
and not expect any further lookups.

`isHtmlExt` is moved to the temporary property of the wrapper fragment
since 42425e6 prematurely stopped keeping it for the wrapped content.

We probably didn't notice this because $wgRawHTML was turned off for
foundationwiki in operations/mediawiki/config commit 2c3e911.

See T201285

Tested on the following wiki where it's still enabled,
http://localhost:8000/donate.wikimedia.org/v3/page/html/WMF-video-2016-noform/27477

The extlink class pass needs to run before cleanup since it accesses
data attributes.

Change-Id: Iafeb329ac96c453d1d012c49989f62dbc98ceee8