Port XMLSerializer.js
The original code used a mimimal amout of the entities npm module;
that has been replaced with the encodeHtmlEntities() method.
Otherwise pretty much 1:1 port.

String length is now calculated in UTF-8 (technically a breaking
change since the JS code used .length which is UCS2; hopefully
OK as long as it is changed consistently everywhere).

Change-Id: I18711372d3d0a7a86d207f6bfd7856103725776a