Update tools/js2php to add codestyle/linter annotations
Run the PHP codestyle and lint tool on the files after they've been
translated from JS, and automatically add annotations to turn off
the tools for specific files if they fail.

This way we implement a ratchet: hopefully those files which pass
codestyle/linting can be kept in such a state.

Change-Id: I4028de946e85ff84ddd5ffc53c4533d047feb779