Allow scalar typehints
mediawiki-codesniffer has a warning about PHP 7+ features that
might not work in HHVM under some configurations. Parsoid does
not aim for HHVM support and requires PHP 7+ so no reason to have
this restriction.

(Filed T216099 about having a separate sniff for the PHP 7.0
compatibility checks which we presumably want to preserve.)

Change-Id: Icf1eb74e683c049b291bfcb431340e887d718eff