Fix `scrubWikitext` mode of normalization in bin/parse.js
And improve its documentation.

This was added in 6ed21f3 but never worked since the option is marked as
a bool.  But this is broken more insidiously in that the Normalizer is
expected to be called on a DOM that has had its data attributes loaded.

Fixing that is a little more daunting so saving it for a follow up.

Change-Id: I7137a703b532cba94e1fa73614e0bbb54de8cc3d