Introduce new DOMUtils.{match,has}TypeOf()/{match,has}NameAndTypeOf() helpers
This splits the typeof attribute and then matches on each element, returning
the matching typeof value.

The existing DOMUtils.isNodeOfType() helper has been renamed to
{match,has}NameAndTypeOf() and tweaked to use DOMUtils.matchTypeOf(),
which magically makes it handle multiple space-separated typeof clauses.
Audited to make sure this is ok.

The `match` version always takes a regular expression and returns the
matching type, and the `has` version always takes a literal string and
returns a boolean.  This eases porting to PHP straight-forward, which
doesn't distinguish between regular expressions and string.

Change-Id: I4521df07ce3be08155c6c241da991fbde85a03b6