PHPUtils: Improve makeSet, deprecate makeMap
PHP has a built-in equivalent to makeSet(), but the wrapper doesn't hurt
anything except a little speed for the extra function call and is
slightly semantically clearer.

But makeMap() is too strange to not deprecate, since PHP has a literal
syntax for associative arrays and the syntax is far better than this
method.

Change-Id: I3c84c3356ac8550b6059353fa821756c2e9f039f