Restrict reinsertable fostered content to internal metas
The only source of fosterable content found after treebuilding should be
the metas that Parsoid itself preserves in those locations.  Restrict
what's tunnelled to those and assert that everything else is allowed
tags, whitespace, and comments.

Use the same mechanism for tunnelling as done during treebuilding.

This is a follow up to I1e9b56ebee4975e7f7642576cd4f06d52e78f26a where a
less precise technique is used.  Keeping that as a separate commit,
rather than squashing, in case the reasoning here is missing something
and we need to revive it.

Change-Id: I204bbd586edccf0138677d60d2e31aee744acc56