Use instanceof for token types
instanceof works, and implemented efficiently: the class_entry*
for the RHS is cached in the opline, so it just has to walk the
inheritance chain for the LHS, which is usually pretty short.

For memory efficiency, remove Token::$type, just get it from reflection
since it is always equal to the class name.

Remove TokenUtils::isOfType() since it is adequately served by
instanceof and is_string(). Leave in getTokenType(), since although it
has no remaining callers, it could conceivably be useful.

Change-Id: I8ba4f9f6a708a975ce1ddc3c05cef1a6820233d9