Add PHPUtils::pushArray()
The considerations are similar to JavaScript. The array has to be passed
by reference explicitly, whereas in JavaScript this is the default. I
omitted the return value to make it more clear that the parameter is
being modified.

Benchmarking shows that it is necessary. Naively using array_merge()
with a small array on the RHS requires copying the LHS. Passing the LHS
by value, appending to it and returning it also requires copying the
LHS. I also benchmarked the case of a small array on the LHS joined to a
large array on the RHS, there was a ~7µs constant overhead.

Change-Id: Ib6d0e2c8c47cbbaf55d1af9bc03a1207aca812d2