Implement TokenUtils::stripEOFTkFromTokens() and placeholder()
In stripEOFTkFromTokens(), pass the array by reference and return
nothing, since that is efficient.

The JS code had a thing to preserve the "rank" property, not sure what
we're going to do about that since PHP arrays can't have properties in
the same way, but the relevant code is unported as yet, and whatever
solution we choose probably won't change this function.

Change-Id: I6fc8fa22c620f8f53993718fe4941185a3fe9204