Fix phan warnings for DOMDataUtils.php
We are deliberately using a dynamic property; add a helper for all
accesses to this property so we can suppress phan's warnings once.

Some actual issues where we needed to use DOMCompat and/or alternative
DOM properties to replace missing PHP APIs.

We still haven't ported the `semver` module, but set it to null to
force a crash if we try to use this code (and to shut up phan).

Change-Id: I4462e8873650346f0b45c670282e3db555f40141