Centralize head/body caching in DOMCompat
A reasonable DOM implementation (aka not PHP's DOM extension) will cache
the head and body in the document, as well as ensure that the caches are
invalidated if the head and/or body are ever modified.

For now we'll do a lazy version of cache in DOMCompat.  Try not to
expose these dynamic properties outside of DOMCompat, though, since
that confounds static analysis and could lead to subtle issues if
someone created a DOMDocument via some mechanism other than
Env::createDocument and was then surprised to find that the dynamic
head/body properties didn't exist on it.

Change-Id: Ib58eebaf41ddb81076e355d96d0c122dfeb419e5