Port Pre and Nowiki extensions
Note for the <nowiki> regexp in Pre: \s in the regular expression is left
as-is. In PHP it is interpreted more narrowly (same as [ \r\n\t\f\v] vs.
[ \f\n\r\t\v\u00a0\u1680\u2000-\u200a\u2028\u2029\u202f\u205f\u3000\ufeff]
in Javascript) but that should be OK - the PHP parser uses
DOMDocument::loadXML and the XML spec defines whitespace as [ \t\r\n].

Change-Id: I8f8a17d099678d8b0343c8a2e46920b3aa0edfb6