Call domino's HTMLParser.insertToken() directly
Removes that conversion code that was added in 37a109a when we moved to
using domino's parser.

Also, drop the namespace test from _att().  For one, the property it
added was ignored in this conversion code since it was added in 37a109a.
However, there's no sense in restoring it since, as explained in
c1bfef5, it was only necessary because we were calling public domain
methods instead of skipping the checks as called for in the HTML5
parsing spec, as we now do.

Bug: T72867
Change-Id: I9f82463e58ca77d024e6d9e259540dad4b1b5f27