Throw exceptions from Grammar.pegphp
A comment on assert() indicates that throwing exceptions from the
tokenizer doesn't work. This is untrue in PHP. So, remove the comment,
and have assert() throw exceptions. Also, replace most instances of
assert() with a new convenience function unreachable(), since that
reflects what most of them were calling it for.

Change-Id: Ic41e519d1dc8bdb6970b4f4c3a65ce3d527fee54