Ensure PHP and JS are consistent wrt allowed entities
The `entities` package we're using on the JavaScript side remaps a bunch
of numeric entities in a way inconsistent with the PHP parser.  We also
allowed a bunch of control characters which PHP didn't.

In the PHP port, we've been using html_entity_decode() which is more
consistent with the legacy parser, with the exception of three oddballs:
\u000C, \u10FFFE and \u10FFFF.  Easy enough to hack around this.

This patch makes the two sides consistent (and matches legacy PHP behavior)
so we don't see false failures in hybrid testing.

Bug: T106578
Bug: T113194
Change-Id: I5d384f9dcf83c4ce85457b55aa3d2d7c8ea67cfc