Fix onSOL buglet in html2wt introduced by refactor
In 4d6aa724555a148aa382d24209a0412e1de2fa4d the code was refactored
and `res.match(regexp)` was swapped around to `regexp.test(res)`.
Unfortunately, `res` is a `ConstrainedText` class, not a `string`,
and `ConstrainedText#match` is a special helper function.  Swap
it back to fix the inadvertent breakage.

Change-Id: Ia6ce1f685185800471dd2be0ee6b0ebcd92461c8