Apply urltext optimizations to JavaScript tokenizer
This applies the optimizations from dc38195590fed2801daf7446f50cf54d65c7c66b
and 0adab30cd67229e241127c9083b185719646363d to the JavaScript version
of the tokenizer.

This just makes it a little easier to debug tokenizer issues by ensuring
the productions we match on JS and PHP are identical.

Change-Id: I0a4a7d18df6f1f1a934e4dc6749409b0c0defdea