Bug fix: not all `tagWidths` were being updated
This is a follow-up to 5718dd39f504fb3f3e644d93c6e206f9cda8ea32:
the TokenUtils::fixupOffsets() method must match the recursion
structure of TokenUtils::collectOffsets() exactly, or else there
can be buried tagWidths which are not properly readjusted.
In this case there was a `<nowiki>` tag inside the `k` part
of an attribute KV, and the tagWidths on that weren't getting fixed
up properly.

Change-Id: Ifd3611f6b0c65eb19f08645a1b6ee786103ac00d