Ensure that proper source texts are used when parsing
It is important to properly match the srcOffsets and source texts so
that we can refer to the original source text in order to adjust the
offsets as needed for different character encodings.

Add a Frame object to the environment which will always hold the
proper source text.  Note that the current implementation needs
to hold/pass the current frame because asynchronous pipelines in
JavaScript are inconsistent with holding current frame in
`Env#currentFrame'.

In order to consistently have srcOffsets to use when invoking recursive
pipelines, we also needed to be much more careful about passing in
appropriate srcOffsets when creating KV pairs for attributes.  This also
allowed us to removed the 'srcOffset' attribute from the DOMFragmentToken
and replace it with proper srcOffset properties on the 'content' KV, which
solves a typing issue with the PHP port of KV.

The transformTests input files were updated to add missing srcOffsets.

Change-Id: Ie25bb638e523def48cab911de6aa6c09e81cdf2c