Add hybrid sanitizer/tokenizer tests to package.json
In addition, the hybrid sanitizer test was added to `composer test` to ensure
we don't regress now that the ported PHP sanitizer has reached parity with
the JS version.

(We can't add the hybrid tests to `npm test` because apparently our
node environment for CI doesn't have PHP installed.  Sigh.)

Change-Id: I80c99b93b6169ee7ffb22e1d9d3d7bfa53aaceea