Fix console.assert when running under node >= 0.10
For browser compatibility (I guess) in Node v10, console.assert() was
changed to log messages to stderr *WITHOUT THROWING AN EXCEPTION*.
That interacts poorly with our service framework, which relies on
exception handling to return error pages, and with various places
in our code which assume that console.assert() can't fall through.

In theory we should probably switch to a "better" assertion library.
But since we don't use console.assert very much anyway, and we're
switching to PHP (with its own assertion library), for the moment
just hack console.assert() to ensure it throws an exception the
way it always used to.

Change-Id: I223370d23448c0ee72c5dbe7c609199056286cff