Fix phan issues found by phan 1.3.2
The latest version of phan throws some lint errors which phan 1.3.1 did not.

These appear to mostly be false positives; tighter types on the variables
involved remove most of them.

SiteConfig seems to show a phan bug, however: the late initialization of
SiteConfig::nsNames in SiteConfig::loadSiteData (and in particular, the
initial `$this->nsNames = []` statement) appears to confuse phan about
the type of the `nsNames` property, completely overriding the type set
by the `@var` annotation at the declaration site.  Moving the
initializations to the constructor works around this problem.

Change-Id: I04b3fb835e190b7092152fb71d1765ca2bffb28d