Back port latest Sanitizer::normalizeCss to Parsoid (JS and PHP)
Copy the latest Sanitizer::normalizeCss from mediawiki core back into
the PHP port.  Back-port this to JS to account for various divergences
over time.  In particular, Parsoid commit
898c9b9ec1443cd6510313e20059e1ec26f04e09 added some mismatched-quote
handling which was never present in core.  We think the need for this
has long past, so drop this divergence on the JS side.

Change-Id: Ide0d1b6f503f17124aa9546dcb857adf1938a453