Start porting DOMPostProcessor
This allows for the construction of a basic post-processor, which runs a
list of processors.  The default list is used if none is provided.

Several PORT-FIXMEs remain, such as adding the extension processors to
the default list and populating the document head with all the meta
information.

Also, a few of the default passes are yet to have been ported so they
remain commented out.

The beginnings of a test suite is added and, in the follow up, the
ported DOMPostProcessor is used in hybrid testing's runDOMPass.

Remove the genTest logic since that's no longer useful post-port.

Bug: T219941
Change-Id: If7c5ffc895aee9d60c6812cba46ea317c30e677d