LinkHandlerUtils: Fix translation of non-boolean JS &&/||
The JS side had

 var sizeString = (size && size.ak) || (ww.fromDataMW && ww.value);

in JS this takes the value of size.ak or ww.value (or something falsey).
The naively translated PHP, on the other hand, always results in true
or false. We need more complicated logic to duplicate the JS behavior.

Further, we may also need to take into account here the fact that in JS
string-0 is truthy while in PHP it's falsey.

Also, apparently Phan was confused all this time about what $getLastOpt
could return, and the changed code structure made it start complaining.
A typehint fixes that.

Change-Id: I26fb19323821584491d93ad83cb3f932b0e289db