Move the Frame class to its own JS file
This was already done in the PHP port.  Backport the change to the JS
port to allow us to move the Frame to the Env#currentFrame.  (Precise
tracking of source texts via the Frame is necessary to allow conversion
of offsets between character encodings in the PHP port and for hybrid
testing.)

Change-Id: I9a5e479b585cbccc51ec289902a883acc5872908