Remove `manager` from JS Frame object
This just syncs the Frame class between JS and PHP ports; `manager`
was already replaced with `env` in the PHP port; this applies that
same change to the JS side.  Removed `this.manager.attributeType`
from implementation of `Frame#expand` -- it was always equal to the
default (`tokens/x-mediawiki`) whenever `Frame#expand` was invoked.

Change-Id: I1ea77a2d462fcf78648a512ff48ba2a81fe8b634